refactor(components): merge useFatchWeather \& useFetchCityWeather
МЕТА ДАННЫЕ: refactor 10
