+++ src/pages/HomePage.vue
@@ -26,7 +26,7 @@
 
 <script setup>
 import TheChosenLocation from '@/components/TheChosenLocation.vue';
-import { useFetchWeather } from '@/hooks/useFetchWeather';
+import useFetchWeather from '@/hooks/useFetchWeather';
 import TheCurrentForecast from '@/components/TheCurrentForecast.vue';
 import TheForecast from '@/components/TheForecats.vue';
 import useFirstUpperCase from '@/hooks/useFirstUpperCase';
